Refactor: request validation is in RequestHelper, user validation in UserHelper:
