+++ app/controllers/api/v1/sessions_controller.rb
@@ -3,11 +3,6 @@ class Api::V1::SessionsController < ApplicationController
 
   def create
     user = User.find_by(email: @body[:email])
-
-    if user && user.authenticate(@body[:password])
-      render json: Api::V1::UserSerializer.user_show(user)
-    else
-      bad_credentials
-    end
+    validate_login(user)
   end
 end
