+++ app/controllers/application_controller.rb
@@ -1,7 +1,4 @@
 class ApplicationController < ActionController::API
-  include ValidationHelper
-
-  def parse_json
-    @body = JSON.parse(request.body.read, symbolize_names: true)
-  end
+  include RequestHelper
+  include UserHelper
 end
