+++ app/modules/request_helper.rb
@@ -1,4 +1,4 @@
-module ValidationHelper
+module RequestHelper
   def validate_content_type
     if request.headers['CONTENT_TYPE'] != 'application/json'
       render json: { error: 'invalid content type' }, status: 400
@@ -11,13 +11,7 @@ module ValidationHelper
     end
   end
 
-  def validate_api_key
-    if !User.valid_key?(@body[:api_key])
-      bad_credentials
-    end
-  end
-
-  def bad_credentials
-    render json: { error: 'bad credentials' }, status: 401
+  def parse_json
+    @body = JSON.parse(request.body.read, symbolize_names: true)
   end
 end
